+++ Source/OptionsComponent.cpp
@@ -1,33 +1,47 @@
 #include "OptionsComponent.h"
+#include "PluginEditor.h"
 
 
 
-
-OptionsComponent::OptionsComponent(std::string ComboTemplate,std::string SliderTemplate,std::string ButtonTemplate)
+OptionsComponent::OptionsComponent(FxseqAudioProcessorEditor *ape,std::string ComboTemplate,std::string SliderTemplate,std::string ImageButtonTemplate)
 {
     setSize (260, 150);
+    comboTemplate=ComboTemplate;
+    sliderTemplate=SliderTemplate;
+    imageButtonTemplate=ImageButtonTemplate;
+    APE=ape; 
+    
+    presetSelected.addItem(TRANS("Default"),1);presetSelected.addItem(TRANS("Demo"),2);
+    presetSelected.setEditableText (true);
+    //for(int i=0;i<comboColors.size();i++) {  presetSelected.setColour(comboColors[i][0],juce::Colour(comboColors[i][1]));   }
+    presetSelected.setSelectedItemIndex(0);
+    presetSelected.onChange = [this] {changeSelectedPreset();};
+    addAndMakeVisible (presetSelected);
+    presetSelectedLabel.setText("Preset",juce::dontSendNotification);
+    addAndMakeVisible (presetSelectedLabel);
 
-    initSlider1("Out gain",outGain,outGainLabel,0.0f,2.0f,0.1f);outGain.setValue(1.0f);
-    initSlider1("Dry/wet",outMix,outMixLabel,0.0f,1.0f,0.1f);outMix.setValue(1.0f);
+    savePreset.setToggleable(true);
+    //for(int i=0;i<imageButtonTemplate.size();i++) {  savePreset.setColour(imageButtonColors[i][0],juce::Colour(imageButtonColors[i][1]));   } // CRASHES THE PLUGIN!!!
+    savePresetLabel.setText("Save",juce::dontSendNotification);
+    addAndMakeVisible (savePreset);
+    addAndMakeVisible (savePresetLabel);
 }   
 
 void OptionsComponent::resized()
 {
-    outGainLabel.setBounds(195,00,80,20);
-    outGain.setBounds     (185,05,80,60);
-    outMixLabel.setBounds (195,58,80,20);
-    outMix.setBounds      (185,63,80,60);
+    presetSelected.setBounds        (10,15,150,20);
+    presetSelectedLabel.setBounds   (50,00,80,20);
+    savePreset.setBounds            (180,15,40,20);
+    savePresetLabel.setBounds       (177,0,40,20);    
 }
 
 void OptionsComponent::skinChange()
-{   /*
-    for(auto templateColor : sliderColors) {
-        outGain.setColour(templateColor[0],juce::Colour(templateColor[1]));
-        outMix.setColour(templateColor[0],juce::Colour(templateColor[1]));
-    }
-    outGainLabel.setColour(0x1000281,juce::Colour(sliderColors[3][1]));
-    outMixLabel.setColour(0x1000281,juce::Colour(sliderColors[3][1]));*/
-    int i;
+{   
+    for(int i=0;i<comboColors.size();i++) {  presetSelected.setColour(comboColors[i][0],juce::Colour(comboColors[i][1]));   }
+    //for(int i=0;i<imageButtonTemplate.size();i++) {  savePreset.setColour(imageButtonColors[i][0],juce::Colour(imageButtonColors[i][1]));   } // CRASHES THE PLUGIN!!!
+
+    presetSelectedLabel.setColour(0x1000281,juce::Colour(comboColors[1][1]));
+    savePresetLabel.setColour(0x1000281,juce::Colour(comboColors[1][1]));
 }
 ////////////////////////////////////////////// TEMPLATES //////////////////////////////////////////////
 void OptionsComponent::initSlider1(std::string name,juce::Slider& slider,juce::Label& label,float min,float max,float def)
@@ -39,4 +53,9 @@ void OptionsComponent::initSlider1(std::string name,juce::Slider& slider,juce::L
 
     addAndMakeVisible(slider);
     addAndMakeVisible(label);
+}
+
+void OptionsComponent::changeSelectedPreset()
+{
+
 }