+++ HISTORY.md
@@ -2,6 +2,17 @@
 
 ## Change records
 
+### v0.02 22092023 Init project
+#### TODO
+- [x] define all processor parameters => **all but effect specific parameters mapped**
+- [ ] write get/Set state information for all processor parameters
+- [x] continue definition of EffectComponent & OptionsComponent
+- [ ] define SequenceSequencerComponent
+- [ ] update processor gainpPatterns while modifying patterns on GUI (**crashes**)
+- [ ] cleaner way to pass colortemplates
+#### BUGS
+- [ ] setting the textbutton label color make it crash (why???)
+
 ### v0.01 21092023 Init project
 #### TODO
 - [ ] define all processor parameters